requirements_Godleib.txt
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
Django==2.2.16
djangorestframework==3.12.4
flake8==4.0.1
idna==3.3
iniconfig==1.1.1
mccabe==0.6.1
packaging==21.2
pluggy==0.13.1
py==1.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
djangorestframework-simplejwt==4.7.2